requirements.txt
Django==1.10.5
django-localflavor==1.4.1
google-api-python-client==1.6.1
httplib2==0.10.3
oauth2client==4.1.2
psycopg2==2.6.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
